code modefied

+++ src/model/transaction/TransModel.js
@@ -8,9 +8,10 @@ export const getTransactionById = (userId) => {
   return TransSchema.find(userId);
 };
 
-export const deleteTransactionByIds = ( idArg, userId,) => {
-  return TransSchema.deleteMany({
-    userId,
-    _id: { $in: idArg },
-  });
-};
+// export const deleteSingleTransaction = (_id) => {
+//   return TransSchema.findById(_id)
+// }
+
+export const deleteTransactionByIds = (userId, idArg) => {
+  return TransSchema.deleteMany({userId, _id: { $in: idArg } })
+}
